D21_neo4j/04_langchain-neo4j-knowledge-graph/requirement.txt
python-dotenv==1.0.0
openai==0.27.4
langchain==0.0.180
neo4j==5.8.0
networkx==3.1
